change paths alias config
