+++ src/features/error-fallback/views/data-view-error-fallback.tsx
@@ -2,7 +2,7 @@ import { Stack, Text } from '@mantine/core';
 import { useDidUpdate } from '@mantine/hooks';
 import { useEffect, useState } from 'react';
 import { FallbackProps } from 'react-error-boundary';
-import { useAppStore } from 'store/app-store';
+import { useAppStore } from '@store/app-store';
 
 export const DataViewErrorFallback = ({ error }: FallbackProps) => {
   const currentView = useAppStore((state) => state.currentView);
