+++ src/hooks/useUploadFilesHandlers.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import { getSupportedMimeType } from 'utils';
+import { getSupportedMimeType } from '@utils/helpers';
 import { useAppContext } from '@features/app-context';
 import { useAppNotifications } from '@components/app-notifications';
