+++ src/main.tsx
@@ -1,4 +1,4 @@
-import App from 'app';
 import ReactDOM from 'react-dom/client';
+import App from './app';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(<App />);
