+++ src/models.ts
@@ -146,3 +146,7 @@ export interface DataBaseModel {
   name: string;
   schemas: SchemaModel[];
 }
+
+export type ErrorModel = {
+  message: string;
+};
