+++ src/pages/main-page/main-page.tsx
@@ -1,12 +1,11 @@
 import { Stack, useMantineColorScheme } from '@mantine/core';
 import { Allotment } from 'allotment';
-import { useAppContext } from 'features/app-context';
+import { useAppContext } from '@features/app-context';
 import { useHotkeys, useLocalStorage } from '@mantine/hooks';
 import { TabsPane } from '@features/tabs-pane';
 import { Spotlight } from '@mantine/spotlight';
-import { useFileHandlers } from 'hooks/useUploadFilesHandlers';
+import { useFileHandlers } from '@hooks/useUploadFilesHandlers';
 import { DataViewer } from '@features/data-viewer';
-import { SpotlightMenu } from '@components/spotlight';
 import { Navbar } from './components';
 
 export const MainPage = () => {
